Implement more attributes

+++ src/Rocksmith2014.DLCProject/Manifest/PhraseIteration.fs
@@ -1,8 +1,8 @@
 ﻿namespace Rocksmith2014.DLCProject.Manifest
 
 type PhraseIteration =
-    { PhraseIndex : int16
-      MaxDifficulty : byte
+    { PhraseIndex : int
+      MaxDifficulty : int
       Name : string
       StartTime : float32
       EndTime : float32 }
